// source/frontend_web/packages/framework/metafox-html-viewer/src/types.tsx
import { ElementType } from 'react';

export interface THtmlViewerNode {
  type: string;
  name: string;
  children?: [THtmlViewerNode];
  next?: THtmlViewerNode;
  prev?: THtmlViewerNode;
  parent?: THtmlViewerNode;
  data?: string;
  attribs: Record<string, any>;
}

export type THtmlViewerTransform = (
  node: THtmlViewerNode,
  index: string
) => React.ReactNode;

export interface HtmlComponentProps {
  html: string;
  decodeEntities?: boolean;
  transform?: THtmlViewerTransform;
  preprocessNodes?: any;
  component?: ElementType;
}